+++ autotraders/ship/fuel.py
@@ -1,7 +1,9 @@
-class Fuel:
-    def __init__(self, current, total):
-        self.current = current
-        self.total = total
+from pydantic import BaseModel
+
+
+class Fuel(BaseModel):
+    current: int
+    total: int
 
     def __str__(self):
         return str(self.current) + "/" + str(self.total)